Atualizaçoes
11/09/2018

+++ Projeto_Final/src/database/Estrutura.sql
@@ -6,7 +6,7 @@ USE auto_pecas;
 CREATE TABLE produtos(
     id int AUTO_INCREMENT PRIMARY KEY,
     categoria         varchar(50),
-    quantidade        float,
+    quantidade        int,
     valor             float,
     descricao         Varchar(150),
     aplicacao         varchar(300),
